# opendatacon version 1.11.0

## 1.11.0
  * New C API — implement custom Ports, Transforms, and UI plugins in C, or any language that can impement C bindings
    * A Go-based Modbus port (GoModbusPort) is included as a working example
  * New port library: **GoModbusPort** — Modbus over TCP or RTU in both client and server roles
    * `"GoModbusClient"` polls a remote Modbus device; `"GoModbusServer"` acts as a Modbus outstation
    * Supports Binaries, Analogs, Controls, Statuses, OctetStrings with per-point scaling, endian, and poll rates
  * DNP3OutstationPort: new `BatchUpdateBuilder` option — accumulates point updates and flushes them as a single batch
    * Useful for reducing DNP3 event noise when many points change together; batch period adapts automatically to event rate
  * DNP3Port:
    * `"ConnectionlessUDP"` mode for fixed UDP links
    * `"UDPListenPort"` can be set to 0 for an ephemeral port
    * Unspecified '"IP"' (0.0.0.0 or ::) for promiscuous UDP
  * DNP3MasterPort:
    * option to periodically re-enable unsolicited responses
    * pass-through received Analog command type when no fixed type is configured
  * JSONPort: configurable `AnalogControlType`
  * Log format strings are now configurable per-sink via config or UI commands, including UTC vs local time
  * New `EverythingInOne` example config combining all major port types in one file
  * Stability fixes for LuaPort/LuaTransform and PyPort shutdown
  * Python standard library packaged correctly on Windows
  * SimPort SQLite playback bugfixes

+++ Code/Plugins/ConsoleUI/tinycon.cpp
@@ -240,7 +240,7 @@ std::string tinyConsole::getLine (int mode = M_LINE, const std::string& delimete
 void tinyConsole::setBuffer (std::string s)
 {
 	buffer.assign(s.begin(),s.end());
-	line_pos = buffer.size();
+	line_pos = static_cast<int>(buffer.size());
 }
 
 void tinyConsole::reprint_prompt_buffer()
@@ -307,14 +307,14 @@ void tinyConsole::run ()
 							buffer.erase(buffer.begin(), buffer.end());
 
 							pos++;
-							if (pos > ((int)history.size() - 1)) pos = history.size() - 1;
+							if (pos > ((int)history.size() - 1)) pos = static_cast<int>(history.size()) - 1;
 
 							// store in buffer
 							for (size_t i = 0; i < history[pos].size(); i++)
 							{
 								buffer.push_back(history[pos][i]);
 							}
-							line_pos = buffer.size();
+							line_pos = static_cast<int>(buffer.size());
 							// output to screen
 							std::cout << history[pos] << std::flush;
 							break;
@@ -353,7 +353,7 @@ void tinyConsole::run ()
 									}
 								}
 							}
-							line_pos = buffer.size();
+							line_pos = static_cast<int>(buffer.size());
 							break;
 						case LEFT_ARROW:
 							// if there are characters to move left over, do so
